requirements.txt
aiofiles==0.5.0
aiohttp==3.6.3
APScheduler==3.6.3     
astroid==2.4.2
async-timeout==3.0.1   
asynctest==0.13.0      
attrs==20.2.0
beautifulsoup4==4.9.3  
better-profanity==0.6.1
bs4==0.0.1
certifi==2020.6.20     
chardet==3.0.4
colorama==0.4.4
coverage==5.3
discord-pretty-help==1.1.2
discord.py==1.5.1
docutils==0.16
html5lib==1.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
libneko==0.24.9
lxml==4.6.1
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.0
psutil==5.7.2
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tzlocal==2.1
urllib3==1.25.11
webencodings==0.5.1
websockets==8.1
wrapt==1.12.1
yarl==1.5.1
git+https://github.com/Rapptz/discord-ext-menus@84caae8038d0d